backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-appconf==1.0.5
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==23.1
django-imagekit==4.1.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.3
pilkit==2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
whitenoise==6.4.0
